Add background reaping of orphaned child processes
When running in proxy mode keepalived runs a check script to make sure
that HAProxy is running.  Under currently unknown circumstances zombie
processes of the check script and of the curl process spawned by the
check script have been observed which get reparented to the controller.
This commit adds a child reaper to the controller to collect these
zombies and prevents them from accumulating over time.  Apart from
this actual problem the reaping of orphaned child processes is a good
idea anyway as the controller has PID 1 and it and its descendants spawn
child processes which might need reaping.

Signed-off-by: Lars Fenneberg <lf@elemental.net>